Refactored: remove fixtures
